stream: move common async iteration logic to base_async_iterator
