+++ lib/_stream_readable.js
@@ -38,7 +38,7 @@ const {
   ERR_METHOD_NOT_IMPLEMENTED,
   ERR_STREAM_UNSHIFT_AFTER_END_EVENT
 } = require('internal/errors').codes;
-const { ReadableAsyncIterator } = require('internal/streams/async_iterator');
+const ReadableAsyncIterator = require('internal/streams/async_iterator');
 const { emitExperimentalWarning } = require('internal/util');
 var StringDecoder;
