+++ lib/readline.js
@@ -34,7 +34,7 @@ const {
 } = require('internal/errors').codes;
 const { debug, inherits } = require('util');
 const { Buffer } = require('buffer');
-const BufferList = require('internal/streams/BufferList');
+const BufferList = require('internal/streams/buffer_list');
 const EventEmitter = require('events');
 const { StringDecoder } = require('string_decoder');
 const { emitExperimentalWarning } = require('internal/util');
